requirements.txt
asgiref===3.6.0
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.2.3
django-environ==0.9.0
django-model-utils==4.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.6
python-dotenv==0.21.0
sqlparse==0.4.3
static3==0.7.0
whitenoise==6.2.0